Created typedoc adapter
